requirements.txt
langchain>=0.1.13
langchain-openai>=0.0.8
openai>=1.14.1
python-dotenv>=1.0.0
fpdf>=2.7.6
langchain-community>=0.0.27
matplotlib>=3.8.0
yfinance>=0.2.36
requests>=2.31.0
